[REFACTOR] project_register.js 분할로 규모 축소 (api, renderer, ui_util, form_util 등으로 분리)
