--- src/main/webapp/resources/js/admin/project_register.js
@@ -1,296 +1,0 @@
-import { DateUtil } from "../util/date_util.js";
-import { UploadUtil } from "../util/upload_util.js";
-
-let selectedProjectId = null;
-let deletedPictureIds = [];
-const previewList = document.getElementById('imagePreviewList');
-const addBtn = previewList.querySelector('.add-btn');
-
-// DOM이 다 로드된 후 이벤트 리스너 붙이기
-document.addEventListener('DOMContentLoaded', () => {
-    const btnLoad = document.querySelector('.btn-load');
-    if (btnLoad) {
-        btnLoad.addEventListener('click', openModal);
-    }
-    const btnRegister = document.querySelector('.register-btn');
-    if (btnRegister) {
-        btnRegister.addEventListener('click', insertProject);
-    }
-    const btnEdit = document.querySelector('.edit-btn');
-    if (btnEdit) {
-        btnEdit.addEventListener('click', updateProject);
-    }
-    const btnClose = document.querySelector('.modal-close');
-    if (btnClose) {
-        btnClose.addEventListener('click', closeModal);
-    }
-
-    const imageInput = document.getElementById('imageInput');
-
-    // 파일 선택 시 이벤트
-    imageInput.addEventListener('change', function(e) {
-        const files = Array.from(e.target.files);
-        
-        files.forEach(file => UploadUtil.uploadImage(file, addImageBox))
-    });
-});
-
-// 사진 선택 시 동작
-function addImageBox(src, savedFileName) {
-    // 미리보기 박스 생성
-    const box = document.createElement('div');
-    box.className = 'image-upload-box';
-    
-    box.innerHTML = `
-        <img src="${src}" alt="preview">
-        <div class="btn-remove">&times;</div>
-        <input type="hidden" name="project_image_names" value="${savedFileName}">
-    `;
-    
-    // 삭제 이벤트 추가
-    box.querySelector('.btn-remove').onclick = function(e) {
-        e.stopPropagation(); // 부모 클릭 방지
-        box.remove();
-        // 별도의 파일 배열을 관리할것
-    };
-    
-    // 추가 버튼 앞에 삽입
-    previewList.insertBefore(box, addBtn);
-};
-
-// 모달 열기 및 데이터 로드
-async function openModal () {
-    const modal = document.getElementById('projectModal');
-    const spinner = document.getElementById('loadingSpinner');
-    const listContainer = document.getElementById('projectListContainer');
-
-    modal.style.display = 'block';
-    
-    // 1. 초기화: 기존 리스트 비우고 스피너 보여주기
-    listContainer.innerHTML = '';
-    spinner.style.display = 'block';
-
-    try {
-        // 2. API 호출
-        const response = await fetch(`${ctx}/api/projects/all`);
-        if (!response.ok) throw new Error('데이터를 불러오는데 실패했습니다.');
-        
-        const projects = await response.json();
-
-        // 3. 데이터 렌더링
-        renderProjectList(projects);
-    } catch (error) {
-        console.error('Error:', error);
-        listContainer.innerHTML = `<li class="project-item" style="color:red; text-align:center;">Cannot Load Data.</li>`;
-    } finally {
-        // 4. 로딩 종료: 스피너 숨기기
-        spinner.style.display = 'none';
-    }
-}
-
-// 모달 닫기
-function closeModal() {
-    document.getElementById('projectModal').style.display = 'none';
-}
-
-// 프로젝트 리스트 렌더링 함수
-function renderProjectList(projects) {
-    const listContainer = document.getElementById('projectListContainer');
-    
-    if (projects.length === 0) {
-        listContainer.innerHTML = '<li class="project-item">등록된 프로젝트가 없습니다.</li>';
-        return;
-    }
-
-    projects.forEach(data => {
-        const li = document.createElement('li');
-        li.className = 'project-item';
-        
-        // 데이터 구조에 맞춰 HTML 생성
-        li.innerHTML = `
-        <strong>${data.projectName} (${data.projectRound}차)</strong><br>
-        <small>ID: ${data.projectId} | 목표액: ${data.targetAmount.toLocaleString()}원 | 예상수익: ${data.expectedReturn}%</small>
-        `;
-        
-        // 클릭 시 데이터 바인딩 이벤트
-        li.onclick = () => selectProject(data);
-        listContainer.appendChild(li);
-    });
-}
-
-
-// 선택한 프로젝트 정보를 폼에 자동 입력 (차수는 +1)
-async function selectProject(data) {
-    selectedProjectId = data.projectId;
-    // JSP 내의 input name/id에 맞춰 매핑
-    document.getElementsByName('farm_id')[0].value = data.farmId;
-    document.getElementsByName('project_name')[0].value = data.projectName;
-    document.getElementsByName('project_round')[0].value = data.projectRound + 1; // 새 차수 등록용
-    document.getElementsByName('project_description')[0].value = data.projectDescription || '';
-    document.getElementsByName('target_amount')[0].value = data.targetAmount;
-    document.getElementsByName('min_amount_per_investor')[0].value = data.minAmountPerInvestor;
-    document.getElementsByName('expected_return')[0].value = data.expectedReturn;
-    document.getElementsByName('manager_count')[0].value = data.managerCount;
-    // DTO의 필드명과 JSP의 input name을 매핑합니다.
-    const dateFields = {
-        'announcement_start_date': data.announcementStartDate,
-        'announcement_end_date': data.announcementEndDate,
-        'subscription_start_date': data.subscriptionStartDate,
-        'subscription_end_date': data.subscriptionEndDate,
-        'result_announcement_date': data.resultAnnouncementDate,
-        'project_start_date': data.projectStartDate,
-        'project_end_date': data.projectEndDate
-    };
-
-    // 각 날짜 필드 순회하며 값 입력
-    for (const [name, value] of Object.entries(dateFields)) {
-        document.getElementsByName(name)[0].value = DateUtil.toLocalDateTime(value);
-    }
-    
-    // 사진 정보 가저오기
-    try {
-        const response = await fetch(`${ctx}/api/project/picture/${data.projectId}/all`);
-        if (!response.ok) throw new Error('사진 정보를 가져오지 못했습니다.');
-        
-        const picturesData = await response.json();
-        if (picturesData) {
-            displayExistingImages(picturesData)
-        }
-    } catch (error) {
-        console.error("Picture API Error:", error);
-        // 사진 정보가 없을 경우 빈칸으로 유지하거나 사용자 알림
-    }
-
-    // 토큰 정보 가져오기
-    try {
-        const response = await fetch(`${ctx}/api/token/${data.projectId}`);
-        if (!response.ok) throw new Error('토큰 정보를 가져오지 못했습니다.');
-        
-        const tokenData = await response.json();
-        
-        if (tokenData) {
-            // JSP의 input name인 token_name과 ticker_symbol에 값 세팅
-            document.getElementsByName('token_name')[0].value = tokenData.tokenName || '';
-            document.getElementsByName('ticker_symbol')[0].value = tokenData.tickerSymbol || '';
-            document.getElementsByName('total_supply')[0].value = tokenData.totalSupply || '';
-        }
-    } catch (error) {
-        console.error("Token API Error:", error);
-        // 토큰 정보가 없을 경우 빈칸으로 유지하거나 사용자 알림
-    }
-
-    closeModal();
-}
-
-// 폼 데이터를 JSON으로 만드는 공통 함수
-function getFormData() {
-    const form = document.querySelector('form');
-    const formData = new FormData(form);
-    const obj = {};
-    
-    const arrayFields = ['project_image_names', 'projectImageNames'];
-    
-    // FormData를 JSON 객체로 변환
-    formData.forEach((value, key) => {
-        const camelKey = key.replace(/_([a-z])/g, (g) => g[1].toUpperCase());
-
-        if (arrayFields.includes(key) || arrayFields.includes(camelKey)) {
-            if (!obj[camelKey]) {
-                obj[camelKey] = [];
-            }
-            obj[camelKey].push(value);
-            return;
-        }
-
-        if (key === 'target_amount' || key === 'project_round') {
-            obj[camelKey] = Number(value);
-        } else if (key.includes('date') && value) {
-            obj[camelKey] = DateUtil.toOffsetDateTime(value);
-        } else {
-            obj[camelKey] = value;
-        }
-    });
-    
-    // 수정일 경우 ID 포함
-    if (selectedProjectId) {
-        obj['projectId'] = selectedProjectId;
-    }
-    obj['deletedPictureIds'] = deletedPictureIds;
-    return obj;
-}
-
-// 프로젝트 수정
-function updateProject() {
-    const data = getFormData();
-
-	fetch(`${ctx}/api/projects/update`, { method: 'POST',
-        headers: { 
-            'Content-Type': 'application/json; charset=UTF-8',
-            'Accept': 'application/json' 
-        },
-        body: JSON.stringify(data) })
-	.then(res => res.text())
-	.then(status => {
-	    if(status === "success") {
-            alert("수정되었습니다.");
-	        window.location.href = `${ctx}/admin/project/new`;
-	    } else {
-            const errorMsg = status;
-            alert("실패: " + errorMsg);
-        }
-	});
-}
-
-function insertProject() {
-    const data = getFormData();
-
-    fetch(`${ctx}/api/projects/insert`, { method: 'POST',
-        headers: {
-            'Content-Type': 'application/json; charset=UTF-8',
-            'Accept': 'application/json'
-        },
-        body: JSON.stringify(data) })
-    .then(res => res.text())
-    .then(status => {
-        if(status === "success") {
-            alert("등록되었습니다.");
-            window.location.href = `${ctx}/admin/project/new`;
-        } else {
-            const errorMsg = status;
-            alert("실패: " + errorMsg);
-        }
-    });
-}
-
-
-// 기존 이미지들을 화면에 그려주는 함수
-function displayExistingImages(picturesData) {
-    const existingBoxes = previewList.querySelectorAll('.image-upload-box:not(.add-btn)');
-    existingBoxes.forEach(box => box.remove());
-    deletedPictureIds = []; // 모달 새로 열 때 초기화
-    picturesData.forEach(pic => {
-        const box = document.createElement('div');
-        box.className = 'image-upload-box existing';
-        
-        // src는 API 경로, 삭제를 위해 pictureId를 데이터 속성에 저장
-        box.innerHTML = `
-            <img src="${ctx}/uploads/projects/${pic.imageUrl}" alt="project image">
-            <div class="btn-remove" data-id="${pic.projectPictureId}">&times;</div>
-        `;
-
-        box.querySelector('.btn-remove').onclick = function(e) {
-            e.stopPropagation();
-            if(confirm("이 사진을 삭제하시겠습니까?")) {
-                markImageAsDeleted(pic.projectPictureId); // ID 저장
-                box.remove();
-            }
-        };
-        previewList.insertBefore(box, addBtn);
-    });
-}
-
-function markImageAsDeleted(pictureId) {
-    if (pictureId && !deletedPictureIds.includes(pictureId)) {
-        deletedPictureIds.push(pictureId);
-    }
-}
