+++ src/main/webapp/resources/js/domain/token/token_api.js
@@ -1,0 +1,6 @@
+// api/token_api.js
+import { http } from "../../api/http_client.js";
+
+export const TokenApi = {
+    getToken: (projectId) => http.get(`${ctx}/api/token/${projectId}`),
+}
