+++ src/main/webapp/resources/js/util/ui_util.js
@@ -1,0 +1,5 @@
+export const UI = {
+    show(id) { document.getElementById(id).style.display = 'block'; },
+    hide(id) { document.getElementById(id).style.display = 'none'; },
+    clear(id) { document.getElementById(id).innerHTML = ''; }
+}
